feat: added raspberry hardware controls tests video streaming
Have used libcamera with a subprocessing

+++ main_vstream_ws_server.py
@@ -93,7 +93,7 @@ class VstreamWsSocketIOServerProcess(Process):
 
         logging.info("[SocketIO] Server starting...")
             
-        socketio.run(app, host=self.host, port=self.port)
+        socketio.run(app, host=self.host, port=self.port, allow_unsafe_werkzeug=True)
 
 
 class VideoTrackProviderProcess(Process):
@@ -101,7 +101,7 @@ class VideoTrackProviderProcess(Process):
     Video Track provider
     """
     
-    def __init__(self, server_url="http://localhost:8000"):
+    def __init__(self, server_url="http://localhost:8000", os: str="others"):
         super().__init__()
         
         self.stop_event = Event()
@@ -109,6 +109,8 @@ class VideoTrackProviderProcess(Process):
         self.server_url = server_url
         
         self.vstream_server = None
+        self.os = os
+        
         self.loop = None
     
     def run(self):
@@ -135,7 +137,8 @@ class VideoTrackProviderProcess(Process):
         self.vstream_server = SocketIoVstreamServer(
             self.server_url,
             namespaces=["/video"],
-            async_event_loop=self.loop
+            async_event_loop=self.loop,
+            os=self.os
         )
         
         signal.signal(signal.SIGTERM, self.handle_shutdown)
@@ -184,16 +187,21 @@ class VideoTrackProviderProcess(Process):
                 
     
 
-def main(host, port, ws_uri):
-    server = VstreamWsSocketIOServerProcess(host=host, port=port)
-    tracker = VideoTrackProviderProcess(server_url=ws_uri)
+def main(host, port, ws_uri, features: list[str], os: str):
+    server = None
+    track_provider = None
     
 
     try:
-        server.start()
-        logging.info("[Main] Server process scheduled to start")
-        tracker.start()
-        logging.info("[Main] Tracker process scheduled to start")
+        if "server" in features:
+            server = VstreamWsSocketIOServerProcess(host=host, port=port)
+            server.start()
+            logging.info("[Main] Server process scheduled to start")
+        
+        if "video" in features:
+            track_provider = VideoTrackProviderProcess(server_url=ws_uri, os=os)
+            track_provider.start()
+            logging.info("[Main] Track_provider process scheduled to start")
         
         logging.info("[Main] Main process running. Press Ctrl+C to stop.")
         while True:
@@ -202,24 +210,24 @@ def main(host, port, ws_uri):
     except KeyboardInterrupt:
         logging.info("\n[Main] KeyboardInterrupt received. Shutting down...")
 
-        server.terminate()
-        tracker.terminate()
-        
-        server.join(timeout=5)
-        tracker.join(timeout=5)
 
-
-        if server.is_alive():
-            logging.warning("[Main] Server Force killing process...")
-            server.kill()
+        if server is not None:
+            server.terminate()
+            server.join(timeout=5)
+            if server.is_alive():
+                logging.warning("[Main] Server Force killing process...")
+                server.kill()
+                
+        if track_provider is not None:
+            track_provider.terminate()
+            track_provider.join(timeout=5)
             
-        if tracker.is_alive():
-            logging.warning("[Main] Tracker Force killing process...")
-            tracker.kill()
+            if track_provider.is_alive():
+                logging.warning("[Main] Track_provider Force killing process...")
+                track_provider.kill()
 
         logging.info("[Main] Clean exit.")
-    except Exception as e:
-        logging.exception("Exception while running")
+        
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -245,10 +253,28 @@ if __name__ == "__main__":
         help="Host (e.g. http://127.0.0.1:8000)"
     )
     
+    parser.add_argument(
+        "--feature",
+        type=str,
+        action='append',
+        nargs="+",
+        choices=["video", "server"],
+        default=["video", "server"],
+        help="Features to activate, video processing, Video server transmitter"
+    )
+    
+    parser.add_argument(
+        "--os",
+        type=str,
+        choices=["raspberry_pi", "other"],
+        default=["other"],
+        help="other uses the VideoCaptaure and raspberry_pi uses the libcamera as background processs"
+    )
+    
     args = parser.parse_args()
     
     
     try: 
-        main(host=args.io_host, port=args.io_port, ws_uri=args.ws_uri)
+        main(host=args.io_host, port=args.io_port, ws_uri=args.ws_uri, features=args.feature , os=args.os)
     except Exception:
-        pass
+        logging.exception("Exception while running")
